feat: Mejoras completas en flujo de clientes y validaciones
CAMBIOS PRINCIPALES:
1. Migraciones y tipos de datos:
   - Corregi todos los tipos de datos de foreign keys (bigint vs integer)
   - Agregue campo tipo_pago (enum) en tabla pagos
   - Actualice referencias a estados para usar codigo en lugar de id

2. Modelo y Controlador:
   - Agregue descuento_manual a validaciones en controller
   - Implemente 4 tipos de pago: completo, parcial, pendiente, mixto
   - Calculos correctos de precios finales con descuentos
   - Estados correctos segun flujos (100 para inscripcion activa, 200-201 para pagos)

3. Validaciones del Formulario:
   - Mejore handleFormSubmit() con validaciones JavaScript por PASO
   - Validar campos requeridos antes de enviar
   - Validar monto_abonado segun tipo_pago
   - Mostrar errores en alertas Bootstrap (no SweetAlert2)
   - Mensajes de error claros y especificos

4. Seeders:
   - Corregi todos los seeders para usar estado.codigo en lugar de estado.id
   - Validaciones correctas de foreign keys

5. API:
   - Endpoint /admin/api/precio-membresia/{id} retorna precios como integers
   - Calculo correcto de descuentos por convenio

CARACTERISTICAS FUNCIONALES:
 Membresias cargadas correctamente desde BD
 Precios mostrados como enteros (sin decimales)
 3 flujos de cliente: solo_cliente, con_membresia, completo
 4 tipos de pago con validaciones dinamicas
 Resumen en PASO 3 con info de PASO 1 y PASO 2
 Descuento manual aplicable en PASO 2
 Validaciones JavaScript antes de submit
 Alertas Bootstrap que funcionan siempre
 Estados correctamente asignados segun flujo

+++ database/migrations/0001_01_02_000004_create_precios_membresias_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('precios_membresias', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('id_membresia');
+            $table->unsignedBigInteger('id_membresia');
             $table->decimal('precio_normal', 10, 2);
             $table->decimal('precio_convenio', 10, 2)->nullable()->comment('NULL si no aplica convenio');
             $table->date('fecha_vigencia_desde');
