+++ database/migrations/0001_01_02_000011_add_role_to_users_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('id_rol')->default(1)->after('email_verified_at');
+            $table->unsignedBigInteger('id_rol')->default(1)->after('email_verified_at');
             $table->foreign('id_rol')->references('id')->on('roles')->onDelete('restrict');
         });
     }
